+++ app/services/cart_user_creator.rb
@@ -1,10 +1,10 @@
 class CartUserCreator
   def self.call(cart, user_params)
     user = if cart.user.nil?
-            new_user_params = (user_params || {}).merge(guest: true)
-            User.create(new_user_params)
+             new_user_params = (user_params || {}).merge(guest: true)
+             User.create(new_user_params)
            else
-            cart.user
+             cart.user
            end
 
     return ServiceResult.success(user) if user.valid?
